Update app.js
Hey I made some changes getting rid of the updateMain function and adding it to the updateIndexPoints function since they are called in the same spot.  Yaa it worked.
I left in the old stuff wacked out Ha.
I needed to go back and really learn this as my friday brain was not rolling with this fully.  But now I think I stand a better shot or set up to attach the weekend chanllenge.
Cheers